Redis "too many clients" error fixed
It appears the issue was that the data fetch job was creating new instances of redis clients on every loop. This eventually lead to a ` ERR max number of clients reached` error.

The solution was to create a `RedisFactory` module that now handles a `client` singleton that is shared globally across the multiple files that are using redis in one way or another.